+++ src/main/java/br/com/wnfa/alurachallenge/resource/ExpenseResouceSwagger.java
@@ -51,8 +51,8 @@ public interface ExpenseResouceSwagger {
 		@ApiImplicitParam(name = "sort", allowMultiple = true, dataType = "string", paramType = "query", value = "Ordenação dos registros") })
 	@ApiOperation("Lista todas as despesas cadastradas de acordo com o ano e mês informado")
 	public ResponseEntity<Page<ExpenseResponseDTO>> findByMesAndAno(
-			@ApiParam(value = "Filtrar o ano da despesa", example = "2030") @PathVariable Long year,
-			@ApiParam(value = "Filtrar o mês da despesa", example = "12") @PathVariable Long month, 
+			@ApiParam(value = "Filtrar o ano da despesa", example = "2030") @PathVariable Integer year,
+			@ApiParam(value = "Filtrar o mês da despesa", example = "12") @PathVariable Integer month, 
 			@ApiIgnore Pageable pageable
 			);
